+++ ChangeLog.md
@@ -1,5 +1,6 @@
 #### v1.0.15
 * Add ability to store user state object in DicomService based classes
+* Add ability to store user state object in DicomClient
 
 #### v1.0.14 (09/05/2012)
 * Fix bug in DicomDatasetReaderObserver handling zero length Specific Character Set elements